+++ app/Filament/Resources/Users/Pages/EditUserPage.php
@@ -21,6 +21,7 @@ class EditUserPage extends EditRecord
     {
         return [
             UserResource::toggleActiveAction(),
+            UserResource::resetPasswordAction(),
         ];
     }